tipp_scrapping/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
beautifulsoup4==4.14.3
boto3==1.42.94
botocore==1.42.94
certifi==2026.4.22
charset-normalizer==3.4.7
click==8.3.3
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.136.0
fastapi-cli==0.0.24
fastapi-cloud-cli==0.17.0
fastar==0.11.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.13
Jinja2==3.1.6
jmespath==1.1.0
lxml==6.1.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
pydantic==2.13.3
pydantic-extra-types==2.11.1
pydantic-settings==2.14.0
pydantic_core==2.46.3
Pygments==2.20.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.26
PyYAML==6.0.3
requests==2.33.1
rich==15.0.0
rich-toolkit==0.19.7
rignore==0.7.6
s3transfer==0.16.1
sentry-sdk==2.58.0
shellingham==1.5.4
six==1.17.0
soupsieve==2.8.3
starlette==1.0.0
typer==0.24.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.46.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
